requirements.txt
flask
itsdangerous
markupsafe
requests
gunicorn
blinker==1.3
cov-core==1.14.0
coverage==3.7.1
pytest-cov==1.8.0
responses==0.3.0